A function for checking the assumptions of IRT.

This function checks the assumptions of IRT. For unidimensionality the DETECT, RATIO, and ASSI statistics are calculated using sirt::conf.detect(). For local dependence, the standardized LD statistic is calculated using mirt::residuals-method() and the correlations are created using ltm::rcor.test(). Relative fit calculates BIC, AIC, and several others using code from mirt::anova-method(). Absolute fit is calculated only for models run with the mirt engine, using mirt::M2(). See the documentation on those functions for more information on calculation.

irt_assume(wizirt_fit, verbose = FALSE)

Arguments

wizirt_fit
An object from fit_wizirt

verbose
Logical. Should the names of the other packages used in this function be printed? Default is FALSE.

Value

A list with tibbles of data in them.

Examples

data("responses")
my_model <- wizirt(data = responses[,-1])
assumptions <- irt_assume(my_model)
print(assumptions, type = 'rel')

#> # A tibble: 8 x 2
#>   stat    values
#>   <chr>    <dbl>
#> 1 log_lik -1026.
#> 2 N          75 
#> 3 n_pars     25 
#> 4 AIC      2101.
#> 5 AICc     2128.
#> 6 BIC      2159.
#> 7 SABIC    2080.
#> 8 HQ       2124.

print(assumptions, type = 'abs')

#>             M2  df         p      RMSEA RMSEA_5   RMSEA_95     SRMSR       TLI
#> stats 303.2375 300 0.4368765 0.01207609       0 0.04630653 0.1190909 0.9524023
#>             CFI
#> stats 0.9524023

print(assumptions, type = 'unid')

#> # A tibble: 5 x 3
#>   Statistics  Value Conclusion                 
#>   <chr>       <dbl> <chr>                      
#> 1 DETECT     -0.62  Essential Unidimensionality
#> 2 ASSI       -0.290 Essential Unidimensionality
#> 3 RATIO      -0.36  Essential Unidimensionality
#> 4 MADCOV100   1.69  NA                         
#> 5 MCOV100    -0.62  NA                         

print(assumptions, type = 'ld')

#> # A tibble: 300 x 5
#>    item_1 item_2 LD_std    pvals   ccov
#>    <chr>  <chr>   <dbl>    <dbl>  <dbl>
#>  1 A      W       0.405 0.000312 0.0241
#>  2 J      P       0.399 0.000398 0.0578
#>  3 K      W       0.363 0.00139  0.0286
#>  4 A      S       0.331 0.00371  0.0512
#>  5 A      K       0.315 0.00592  0.0255
#>  6 O      T       0.309 0.00695  0.0338
#>  7 D      I       0.303 0.00816  0.0382
#>  8 H      W       0.295 0.0101   0.0128
#>  9 M      Q       0.285 0.0132   0.0456
#> 10 H      P       0.282 0.0142   0.0243
#> # ... with 290 more rows