A function used to get additional person-level information, such as person-level fit measures and person response functions. I accidentally output plots in addition to data in one part. That will be moved to a plot method later.

If there is missing data present, non-parametric imputation will be done to get the cutoff for the measures.

irt_person_fit(wizirt_fit, stats = c("Ht"), items = NULL)

Arguments

wizirt_fit
An object coming from the fit_wizirt function.

stats
A character or character string identifying person-level fit measures. Default is "Ht". All of the stats in PerFit should be available. Let me know if any don't work. for more information.

items
Which items to plot? Either a numeric vector of item positions in the column names of the data, or a vector of the item names to include. If nothing is specified all items are included.

Value

A list with person-level statistics, person-response functions, data for person-response functions, and an empty slot for multi-level information that will be coming soon.

Examples

pfa <- wizirt2:::irt_person_fit(my_model)

#> Error in is.data.frame(x): object 'my_model' not found