Create a general IRT report for a test

irt_report

irt_report(
  data,
  report_type = "html",
  title = "My Report",
  author = "",
  note = "",
  engine = "mirt",
  item_type = "Rasch",
  rownames = NULL,
  tol = 1e-05,
  irt_pars = TRUE,
  ifa_stats = c("X2"),
  pfa_predictors = NULL,
  pfa_bins = 5,
  pfa_stats = c("Ht")
)

Arguments

item_type
Character. Must be one of 'Rasch', '1PL', '2PL' or '3PL'.

rownames
Optional unique row IDs for the data (i.e. examinee IDs). If omitted, uses 1:nrow(data).

tol
Numeric. Convergence criterion. Currently only implemented when engine is mirt.

irt_pars
Logical. Should the traditional IRT parametrization be used? If false the slope-intercept form is used.

ifa_stats
A character or character string identifying item-level fit measures. Must be at least one of c('Zh', 'X2', 'G2', 'infit'). More are coming very soon. Default is 'X2'.

pfa_predictors
Either a matrix or dataframe with a row per person and a column per predictor.

pfa_bins
Numeric. How many bins to break items into for pfa multilevel modeling. Sorted by difficulty then broken into bins. If items cannot fit equally into bins, the bins of easier items will take extras.

pfa_stats
A character or character string identifying person-level fit measures. Default is "Ht". All of the stats in PerFit should be available. Let me know if any don't work. for more information.

mode
Must be 'regression' currently.